add a confirm for pupate

+++ code/modules/antagonists/brain_slug/abilities/pupate.dm
@@ -13,15 +13,19 @@
 		if (!H.slug)
 			boutput(holder.owner, "<span class='notice'>You do not have a brain slug inside of you!</span>")
 			return TRUE
-		H.make_jittery(1000)
-		H.emote("scream")
-		H.visible_message("<span class='alert'>[H]'s flesh appears to contort and shift as if something was trying to come out of [him_or_her(H)]!</span>", "<span class='alert'>You begin to evolve into your final form.</span>")
-		H.changeStatus("slowed", 10 SECONDS)
-		SPAWN (10 SECONDS)
-			if (!H || !H.slug)	return TRUE
-			var/mob/living/critter/adult_brain_slug/new_slug = new /mob/living/critter/adult_brain_slug(H)
-			H.mind.transfer_to(new_slug)
-			new_slug.visible_message("<span class='alert'>A gigantic monster bursts out of [H]! [pick("GOOD LORD!", "GOD JESUS WHAT THE FUCK!", "You're going to be sick")]!</span>", "<span class='notice'>You rise!</span>")
-			var/mob/living/critter/brain_slug/old_slug = H.slug
-			qdel(old_slug)
-			H.gib()
+		var/choice = tgui_alert(holder.owner, "Are you sure you wish to evolve?", "Evolution", list("Yes", "No"))
+		if (!choice || choice == "No")
+			return TRUE
+		if (choice == "Yes")
+			H.make_jittery(1000)
+			H.emote("scream")
+			H.visible_message("<span class='alert'>[H]'s flesh appears to contort and shift as if something was trying to come out of [him_or_her(H)]!</span>", "<span class='alert'>You begin to evolve into your final form.</span>")
+			H.changeStatus("slowed", 8 SECONDS)
+			SPAWN (8 SECONDS)
+				if (!H || !H.slug)	return TRUE
+				var/mob/living/critter/adult_brain_slug/new_slug = new /mob/living/critter/adult_brain_slug(H)
+				H.mind.transfer_to(new_slug)
+				new_slug.visible_message("<span class='alert'>A gigantic monster bursts out of [H]! [pick("GOOD LORD!", "GOD JESUS WHAT THE FUCK!", "You're going to be sick")]!</span>", "<span class='notice'>You rise!</span>")
+				var/mob/living/critter/brain_slug/old_slug = H.slug
+				qdel(old_slug)
+				H.gib()
